+++ src/main/java/baseball/domain/Score.java
@@ -1,0 +1,19 @@
+package baseball.domain;
+
+public class Score {
+    private int strike;
+    private int ball;
+
+    public Score(int strike, int ball) {
+        this.strike = strike;
+        this.ball = ball;
+    }
+
+    public int getStrike() {
+        return strike;
+    }
+
+    public int getBall() {
+        return ball;
+    }
+}
